♻️ refactor: refactor the ChatMessage type
