+++ src/features/Conversation/ChatList/Extras/type.ts
@@ -1,5 +1,5 @@
 import { FC } from 'react';
 
-import { ChatMessage } from '@/types/chatMessage';
+import { ChatMessage } from '@/types/message';
 
 export type RenderMessageExtra = FC<ChatMessage>;
