+++ src/services/__tests__/chat.test.ts
@@ -6,7 +6,7 @@ import { VISION_MODEL_WHITE_LIST } from '@/const/llm';
 import { DEFAULT_AGENT_CONFIG } from '@/const/settings';
 import { useFileStore } from '@/store/file';
 import { useToolStore } from '@/store/tool';
-import { ChatMessage } from '@/types/chatMessage';
+import { ChatMessage } from '@/types/message';
 import { OpenAIChatStreamPayload } from '@/types/openai/chat';
 import { LobeTool } from '@/types/tool';
