+++ src/services/config.ts
@@ -5,8 +5,8 @@ import { useGlobalStore } from '@/store/global';
 import { settingsSelectors } from '@/store/global/selectors';
 import { useSessionStore } from '@/store/session';
 import { sessionExportSelectors, sessionSelectors } from '@/store/session/selectors';
-import { ChatMessage } from '@/types/chatMessage';
 import { ConfigFile } from '@/types/exportConfig';
+import { ChatMessage } from '@/types/message';
 import { LobeSessions } from '@/types/session';
 import { GlobalSettings } from '@/types/settings';
 import { ChatTopic } from '@/types/topic';
