+++ src/store/chat/helpers.ts
@@ -1,5 +1,5 @@
 import { LobeAgentConfig } from '@/types/agent';
-import { ChatMessage } from '@/types/chatMessage';
+import { ChatMessage } from '@/types/message';
 import { OpenAIChatMessage } from '@/types/openai/chat';
 import { encodeAsync } from '@/utils/tokenizer';
