+++ src/store/chat/slices/message/action.test.ts
@@ -7,7 +7,7 @@ import { chatService } from '@/services/chat';
 import { messageService } from '@/services/message';
 import { topicService } from '@/services/topic';
 import { chatSelectors } from '@/store/chat/selectors';
-import { ChatMessage } from '@/types/chatMessage';
+import { ChatMessage } from '@/types/message';
 
 import { useChatStore } from '../../store';
