+++ src/store/chat/slices/message/reducer.test.ts
@@ -1,4 +1,4 @@
-import { ChatMessage } from '@/types/chatMessage';
+import { ChatMessage } from '@/types/message';
 
 import { MessageDispatch, messagesReducer } from './reducer';
