+++ src/store/chat/slices/message/reducer.ts
@@ -1,7 +1,7 @@
 import isEqual from 'fast-deep-equal';
 import { produce } from 'immer';
 
-import { ChatMessage } from '@/types/chatMessage';
+import { ChatMessage } from '@/types/message';
 import { merge } from '@/utils/merge';
 
 interface UpdateMessage {
