+++ src/store/chat/slices/message/selectors.test.ts
@@ -5,7 +5,7 @@ import { INBOX_SESSION_ID } from '@/const/session';
 import { ChatStore } from '@/store/chat';
 import { initialState } from '@/store/chat/initialState';
 import { useSessionStore } from '@/store/session';
-import { ChatMessage } from '@/types/chatMessage';
+import { ChatMessage } from '@/types/message';
 import { MetaData } from '@/types/meta';
 import { merge } from '@/utils/merge';
