+++ src/store/chat/slices/topic/action.test.ts
@@ -6,7 +6,7 @@ import { LOADING_FLAT } from '@/const/message';
 import { chatService } from '@/services/chat';
 import { messageService } from '@/services/message';
 import { topicService } from '@/services/topic';
-import { ChatMessage } from '@/types/chatMessage';
+import { ChatMessage } from '@/types/message';
 import { ChatTopic } from '@/types/topic';
 
 import { useChatStore } from '../../store';
